# frontend/packages/api-client/openapi/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin application server"
    }
  ],
  "paths": {
...
    "/api/v1/assets/{assetId}/values": {
      "put": {
        "tags": ["asset-controller"],
        "operationId": "setValues",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/AssetCustomFieldValueRequest"
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/purchase-date": {
      "put": {
        "tags": ["asset-controller"],
        "operationId": "changePurchaseDate",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SetAssetPurchaseDateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/placement": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "placement",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPlacementResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["asset-controller"],
        "operationId": "move",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssetPlacementRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetPlacementResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/name": {
      "put": {
        "tags": ["asset-controller"],
        "operationId": "rename_1",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RenameAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/lifecycle": {
      "put": {
        "tags": ["asset-controller"],
        "operationId": "changeLifecycleState",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAssetLifecycleStateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/condition": {
      "put": {
        "tags": ["asset-controller"],
        "operationId": "changeCondition",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangeAssetConditionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/locations/{locationId}/restore": {
      "post": {
        "tags": ["location-controller"],
        "operationId": "restore",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/locations/{locationId}/archive": {
      "post": {
        "tags": ["location-controller"],
        "operationId": "archive",
        "parameters": [
          {
            "name": "locationId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LocationVersionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/categories/{categoryId}/restore": {
      "post": {
        "tags": ["category-controller"],
        "operationId": "restore_1",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/categories/{categoryId}/archive": {
      "post": {
        "tags": ["category-controller"],
        "operationId": "archive_1",
        "parameters": [
          {
            "name": "categoryId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/restore": {
      "post": {
        "tags": ["asset-controller"],
        "operationId": "restore_2",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/assets/{assetId}/archive": {
      "post": {
        "tags": ["asset-controller"],
        "operationId": "archive_2",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/asset-models/{assetModelId}/restore": {
      "post": {
        "tags": ["asset-model-controller"],
        "operationId": "restore_3",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/asset-models/{assetModelId}/custom-fields/{fieldId}/restore": {
      "post": {
        "tags": ["model-custom-field-controller"],
        "operationId": "restore_4",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/asset-models/{assetModelId}/custom-fields/{fieldId}/options/{optionId}/restore": {
      "post": {
        "tags": ["model-custom-field-option-controller"],
        "operationId": "restore_5",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "optionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/custom-fields/{fieldId}/options/{optionId}/archive": {
      "post": {
        "tags": ["model-custom-field-option-controller"],
        "operationId": "archive_3",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "optionId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/custom-fields/{fieldId}/archive": {
      "post": {
        "tags": ["model-custom-field-controller"],
        "operationId": "archive_4",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "fieldId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/asset-models/{assetModelId}/assets": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "list_7",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "includeInactive",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["asset-controller"],
        "operationId": "create_7",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAssetRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/assets/bulk": {
      "post": {
        "tags": ["asset-controller"],
        "operationId": "createBulk",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkCreateAssetsRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/asset-models/{assetModelId}/archive": {
      "post": {
        "tags": ["asset-model-controller"],
        "operationId": "archive_5",
        "parameters": [
          {
            "name": "assetModelId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/v1/assets/{assetId}": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "get_3",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/history": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "history",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetStateChangeResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/assets/{assetId}/contents": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "contents",
        "parameters": [
          {
            "name": "assetId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssetPlacementResponse"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/assets/by-code/{rawCode}": {
      "get": {
        "tags": ["asset-controller"],
        "operationId": "getByCode",
        "parameters": [
          {
            "name": "rawCode",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AssetResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssetCustomFieldValueRequest": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "stringValue": {
            "type": "string"
          },
          "dateValue": {
            "type": "string",
            "format": "date"
          },
          "optionId": {
            "type": "string",
            "format": "uuid"
          }
        },
        "required": ["fieldId"]
      },
      "AssetCustomFieldValueResponse": {
        "type": "object",
        "properties": {
          "fieldId": {
            "type": "string",
            "format": "uuid"
          },
          "fieldName": {
            "type": "string"
          },
          "dataType": {
            "type": "string",
            "enum": ["STRING", "DROPDOWN", "DATE"]
          },
          "displayOrder": {
            "type": "integer",
            "format": "int32"
          },
          "stringValue": {
            "type": "string"
          },
          "dateValue": {
            "type": "string",
            "format": "date"
          },
          "optionId": {
            "type": "string",
            "format": "uuid"
          },
          "optionValue": {
            "type": "string"
          }
        }
      },
      "AssetResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "assetModelId": {
            "type": "string",
            "format": "uuid"
          },
          "assetModelName": {
            "type": "string"
          },
          "publicCode": {
            "type": "string"
          },
          "unitNumber": {
            "type": "integer",
            "format": "int32"
          },
          "individualName": {
            "type": "string"
          },
          "displayName": {
            "type": "string"
          },
          "condition": {
            "type": "string",
            "enum": ["GOOD", "DAMAGED"]
          },
          "lifecycleState": {
            "type": "string",
            "enum": ["ACTIVE", "LOST", "DESTROYED", "RETIRED"]
          },
          "purchaseDate": {
            "type": "string",
            "format": "date"
          },
          "archived": {
            "type": "boolean"
          },
          "metadataIncomplete": {
            "type": "boolean"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetCustomFieldValueResponse"
            }
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "SetAssetPurchaseDateRequest": {
        "type": "object",
        "properties": {
          "purchaseDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
      "AssetPlacementRequest": {
        "type": "object",
        "properties": {
          "locationId": {
            "type": "string",
            "format": "uuid"
          },
          "parentContainerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "AssetPlacementResponse": {
        "type": "object",
        "properties": {
          "assetId": {
            "type": "string",
            "format": "uuid"
          },
          "directLocationId": {
            "type": "string",
            "format": "uuid"
          },
          "parentContainerAssetId": {
            "type": "string",
            "format": "uuid"
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "effectivePath": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "effectivePathText": {
            "type": "string"
          }
        }
      },
      "RenameAssetRequest": {
        "type": "object",
        "properties": {
          "individualName": {
            "type": "string"
          }
        }
      },
      "ChangeAssetLifecycleStateRequest": {
        "type": "object",
        "properties": {
          "lifecycleState": {
            "type": "string",
            "enum": ["ACTIVE", "LOST", "DESTROYED", "RETIRED"]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": ["lifecycleState"]
      },
      "ChangeAssetConditionRequest": {
        "type": "object",
        "properties": {
          "condition": {
            "type": "string",
            "enum": ["GOOD", "DAMAGED"]
          },
          "reason": {
            "type": "string"
          }
        },
        "required": ["condition"]
      },
...
      "LocationVersionRequest": {
        "type": "object",
        "properties": {
          "expectedVersion": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "CreateAssetRequest": {
        "type": "object",
        "properties": {
          "individualName": {
            "type": "string"
          },
          "purchaseDate": {
            "type": "string",
            "format": "date"
          },
          "values": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AssetCustomFieldValueRequest"
            }
          }
        }
      },
      "BulkCreateAssetsRequest": {
        "type": "object",
        "properties": {
          "count": {
            "type": "integer",
            "format": "int32",
            "minimum": 1
          },
          "purchaseDate": {
            "type": "string",
            "format": "date"
          }
        }
      },
...
      "AssetStateChangeResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "changeType": {
            "type": "string",
            "enum": ["CONDITION", "LIFECYCLE"]
          },
          "previousValue": {
            "type": "string"
          },
          "newValue": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "actorUserId": {
            "type": "string",
            "format": "uuid"
          },
          "changedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    }
  }
}